👨‍💻 proof-of-work that downloading of qr works with session dictionary
I spend a lot of time to find  out how to make the qr-png downloadable without storing the png on the server. The function download_qr() did the trick, but it was not so easy to call it from qr(). I finally used a session dictionary to pass over url and error.

In the next commit I want to pass all qr-code parameters.